Add triggered rendering using a `Condvar`
To minimize wasted render cycles, a `Condvar` has been added that is
only triggered when either the application audio engine state has
changed, or when a user input has been received. This allows for evented
rendering to display changes nearly as quickly as they occur.

To prevent lockups in the render loop and when consuming user inputs, a
timeout has been added to the wait condition such that the UI will run
at ~60fps. While this works for now, fast streams of characters are
processed frame-by-frame, rather than being batched, so the input loop
should be modified to read continuously until the iterator blocks before
triggering the render event.