Remove dsp-chain dependency
A major flaw in the `dsp-chain` crate is that it by default sums the
input streams for a given node, meaning that it is impossible to use the
input of one node for modulating parameters.

A new graph structure has been created that allows for multiple input
streams to a given node, and each node must produce a single output stream.