Add gate for sequencing
A primitive gate node has been added that feeds into the ADSR node.
Currently, it's just being used as an on/off trigger on each step, but
it proves the concept and displays that the tempo and other settings can
be modified dynamically while the gate is triggered into the near-future
(i.e. for the number of buffer samples set).

Although working, a mechanism must be added to clear or reset the
scheduled gate triggers such that they will not persist if the audio
engine is stopped and restarted at a previous point in the stream.