Add play/pause transport
Interestingly, this has displayed the reasoning (I think) behind
something I've been curious about in many DAW's over the years. Upon
startup or when changing the audio-engine settings such that it will
"restart", you'll often see the levels spike as if there was a phantom
trigger on your track. This appears to be due to how ADSR's are
initialized - at least in my case it is.

The digital ADSR stores the frame state for gate open and close - upon
initializing we set both the open and close to 0. This works fine since
the frame index is monotonically increasing, but upon initialization,
this can essentially simulate a zero-length trigger on the ADSR, causing
the attack, decay, and release phases to run. This should be relatively
easily solvable by checking if the open frame and close frame are equal,
and in that case, the ADSR can bail early and just return silence (since
this should only ever happen upon initialization).