Add sustain and release cycles to ADSR
After poking deeper into the ADSR workings, there's definitely more than
meets the eye - specifically with regards to rapid triggers, and gates
(triggers) that are open for a shorter period than the attack or decay.
This raises the question of what the behaviour of the ADSR should be -
what is expected, and what is desired.

For this implementation, the ADSR operates on relative levels i.e. from
0-1, so (re)triggering can take a few forms:

* when triggered, drop to 0, then start the ADSR cycle
* when retriggered, drop to 0 then start the ADSR cycle
* when retriggered, restart the ADSR cycle at the current level
* when retriggered, restart the ADSR cycle at the current level, using
  the original attack _slope_

These changes implement the behaviour where the retrigger starts the
ADSR cycle with the current level, maintaining the attack time. With
short attack times, this should have a fairly unnoticeable effect, but
for longer attacks, it may become more evident and create some
interesting audible pattterns.

In addition to retrigger considerations, the gate-close timing must be
considered - i.e. when the gate closes before either the attack or decay
cycles complete. Some brief online research seems to point to most
eurorack ADSR's runnign a full cycle when triggered, so I have followed
suit here and when encountering a 1-frame trigger, the ADSR will simply
become an ADR.

Like with most things in this codebase, I really just don't know and am
making it up as I go along.