Add tabs state struct
The `TabsState` struct can be used for navigating a tab panel using
indices. This may require a change in the future if rather than strings,
an enum is used to specify different views, however the changes required
should be minor, or alternatively a string impl can be used for the
enums used.