Fix playback cursor display position
The playback step sequence is zero indexed, however the display cursor
was being displayed a step ahead of where the true playback is. This
resulted in the appearance "late" playback, as the playhead would be on
step n+1 when the sound for step n is played.

The fix _does_ resolve the issue, however in certain situations, there
will be a slight jump of the playhead when resuming play, as the fix
involves stepping back by one step before continuing. At higher tempos,
this is less noticable, as the step interval is shorter.